This commit enables the reko operators to save much larger datasets to the dataplane. Without these changes the reko operators would timeout and fail to save large quantities of paged reko results.
Details:

In order to avoid Lambda timeouts, persist 10 pages then pass the pagination token to the workflow metadata so the next time the step function "get status" stage restarts this Lambda it can use that token to continue reading reko results from where it left off.

Also in this commit I've removed a lot of duplicated error checks that made the operator code difficult to read.