Add page token to stepfunc (#125)
This commit includes changes necessary to support 2 hour long videos. 

The key to achieving this was to allow step functions to pass a pagination token to the Lambda functions which get results from Rekognition jobs. Now, those Reko operators will persist 10 pages at a time, then stop and pass the pagination token to the step function so it can repeatedly restart the operator Lambda until there are no more pages left to read.

This enables the Reko operators to save much larger datasets to the data-plane. Prior to this commit, Reko operators will timeout when trying to save large quantities of paged Reko result,  which was often the case with label_detection and face_detection.

Other changes:

* Increase lambda timeout and memory in order to help manage requests by Rekognition operators to save large result sets.
* Remove unusued GUI artifacts for Polly and AutoML.
* add API documentation to README
* Split input text in the translate operator so it does not exceed the 5000 characters max allowed by AWS Translate service limit.
* Set timeouts and memory allocations for Lambda functions based on test results from a 2 hour movie (Amélie).
* split bulk elasticsearch inserts in order to avoid exceeding max payload size
* If data is empty, skip ES insert. Data is often empty for operators like content moderation when processing non-explicit videos.

GUI changes:
* Raise max file upload size to 2GB in the GUI.
* allow analysis button to open in new tab
* fade delete alert after 5 seconds
* Change workflow configuration form so users only have to set the language for Transcribe and Translate once. Used to be that users would have to set that language preference twice, but now, since both Transcribe and Translate use the same source language, users can just specify this option once.

CloudFormation template changes:

* Make the Cloudfront URL a clickable link in the outputs from both the webapp CF template and the base stack template.
* Update the email template for the Cognito invite message so it includes a link to the stack.