Remove String macro
Macros must be unique in the AWS account and don't support !Ref or !Sub,
making it harder to deploy multiple instances of MIE in the same AWS
account.

Since the String macro was only being used to create the bucket name,
this commit changes the name to be automatically set, which actually
already uses the StackName.