Do full table scan with query() instead of scan().
Background:

The Dataplane API for listing asset locks requires a full table scan of the LockIndex GSI. It was proposed during code review that we demonstrate this using query() instead of scan(). So, this commit includes the changes needed in order to do that.

Note, scan() would be more efficient than query() since the query has to evaluate the KeyConditionExpression for every row but we've opted to use query in order to predispose software developers looking here for code samples, to use query() instead of scan(), since query() is generally more efficient than scan().