Get source language from translate operator if transcribe operator is disabled.
Background:
This is needed by the Content Localization application. When a user clicks Save Edits to update transcripts, the resulting workflow only runs operators downstream from Transcribe. In that case, the Transcribe operator is disabled, and the webcaptions operators will fail to find the source language if they look for it in the input configuration for the transcribe operator.