SonarQube refactoring pre-work
Background
==========

As part of quality bar initiatives we are scanning our solutions with
[SonarQube](https://sonarqube.org) and refactoring the code to based on
issues raised by the scan.

Change
======

This commit contains minor changes to prepare for some refactoring that
will be done in the next commit. It mainly only renames a couple of
source files and is done in a separate commit only to help maintain git
history across the rename since extensive refactoring will take place.